requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
sqlalchemy==2.0.48
alembic==1.18.4
asyncpg==0.31.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
PyJWT==2.12.1
passlib==1.7.4
bcrypt==5.0.0
python-multipart==0.0.22
aiofiles==25.1.0
pillow==12.1.1
python-slugify==8.0.4
email-validator==2.3.0
httpx==0.28.1